update: 修正WorkerPool构造函数以支持最小工作线程数并优化工作线程管理

+++ utils/WorkerPool.ts
@@ -4,10 +4,15 @@ export default class WorkerPool {
     constructor(workerConstructor: new () => Worker, option?: {
         MaxTaskPerWorker?: number;
         MaxWorkerCount?: number;
+        MinWorkerCount?: number;
     }) {
         this.workerConstructor = workerConstructor;
-        this.MaxTaskPerWorker = option?.MaxTaskPerWorker ?? 5;
+        this.MaxTaskPerWorker = option?.MaxTaskPerWorker ?? 3;
         this.MaxWorkerCount = option?.MaxWorkerCount ?? Infinity;
+        this.MinWorkerCount = option?.MinWorkerCount ?? 0;
+        new Array(this.MinWorkerCount).fill(null).forEach(() => {
+            this.newWorker();
+        });
     }
 
     private workerConstructor: new () => Worker;
@@ -17,35 +22,14 @@ export default class WorkerPool {
 
     private MaxTaskPerWorker: number;
     private MaxWorkerCount: number;
+    private MinWorkerCount: number;
 
-    private get currTaskCount() {
-        return this.workerTaskMap.size;
-    }
-
-    private newWorker() {
-        if (this.workerPool.length >= this.MaxWorkerCount) {
-            console.warn('Worker pool is full, cannot create new worker');
-            return;
-        }
-        const newWorker = new this.workerConstructor();
-        newWorker.onmessage = (event) => {
-            const { id } = event.data;
-            const task = this.workerTaskMap.get(id);
-            if (task) {
-                task.resolve(event.data);
-                this.workerTaskMap.delete(id);
-                this.workerTaskCountMap.set(newWorker, this.workerTaskCountMap.get(newWorker)! - 1);
-                this.terminateInactivityWorker();
-            } else {
-                console.warn(`No task found for worker message with id: ${id}`);
-            }
-        };
-        this.workerPool.push(newWorker);
-        this.workerTaskCountMap.set(newWorker, 0);
+    private get isWorkerAllBusy() {
+        return this.workerPool.every(worker => this.workerTaskCountMap.get(worker)! >= this.MaxTaskPerWorker);
     }
 
     private getInactivityWorker() {
-        if (this.workerPool.length === 0 || this.currTaskCount % this.MaxTaskPerWorker === 0) {
+        if (this.workerPool.length === 0 || this.isWorkerAllBusy === true) {
             this.newWorker();
         }
         // 选择任务数量未达到上限且任务数最多的Worker执行，方便后续对不活跃Worker进行清理
@@ -58,12 +42,40 @@ export default class WorkerPool {
         return sortedWorkers.find(worker => this.workerTaskCountMap.get(worker)! < this.MaxTaskPerWorker) ?? sortedWorkers[sortedWorkers.length - 1];
     }
 
+    private newWorker() {
+        if (this.workerPool.length >= this.MaxWorkerCount) {
+            return;
+        }
+        const newWorker = new this.workerConstructor();
+        newWorker.onmessage = (event) => {
+            const { id } = event.data;
+            const task = this.workerTaskMap.get(id);
+            if (task) {
+                task.resolve(event.data);
+                this.workerTaskMap.delete(id);
+                this.workerTaskCountMap.set(newWorker, this.workerTaskCountMap.get(newWorker)! - 1);
+                this.terminateInactivityWorker();
+            } else {
+                console.error(`WorkerPool: No task found for worker message with id: ${id}`);
+            }
+        };
+        this.workerPool.push(newWorker);
+        this.workerTaskCountMap.set(newWorker, 0);
+
+        // console.log(`WorkerPool: New worker created, current worker count: ${this.workerPool.length}`);
+    }
+
     private terminateInactivityWorker = throttle(() => {
         this.workerPool.filter(w => this.workerTaskCountMap.get(w) === 0).forEach(worker => {
+            if (this.workerPool.length <= this.MinWorkerCount) {
+                return;
+            }
             worker.terminate();
             this.workerPool = this.workerPool.filter(w => w !== worker);
             this.workerTaskCountMap.delete(worker);
         });
+
+        // console.log(`WorkerPool: Inactivity workers terminated, current worker count: ${this.workerPool.length}`);
     }, 1000, {
         trailing: true,
     });
